Make UnwatchedUpdaterService into UnwatchedUpdateWorker.
Move from JobIntentService to WorkManager using a CoroutineWorker, this
should be much more light-weight and not prone to crashes of
JobIntentService.

WorkManager includes dependency on new Guava listenablefuture lib which
conflicts with full Guava version required by API client. So update to
full Guava 27.x which avoids the duplicate class.